Conflicts:
	Unity/Assets/Resources/SpellData.xml

+++ Unity/Assets/Scripts/NPC/gvmNPCData.cs
@@ -3,6 +3,6 @@ using System.Collections;
 
 public class gvmNPCData : MonoBehaviour {
     // currently useless
-    public int HP = 100;
+    public float HP = 100;
     public int state = 0;
 }
